# hs-ontology-api-spec.yaml
openapi: 3.0.3
info:
  title: HubMAP/SenNet Ontology API (hs-ontology-api)
  description: The HuBMAP/SenNet Ontology API contains endpoints for querying a [UBKG](https://ubkg.docs.xconsortia.org/) instance with content from the [HuBMAP/SenNet context](https://ubkg.docs.xconsortia.org/contexts/#hubmapsennet-context). The hs-ontology-api imports the [ubkg-api](https://smart-api.info/ui/96e5b5c0b0efeef5b93ea98ac2794837), which encapsulates both basic connectivity to a UBKG instance and generic endpoint code.
  version: 1.4.4
  contact:
    name: GitHub repository
    url: https://github.com/x-atlas-consortia/hs-ontology-api
servers:
  - url: https://ontology.api.hubmapconsortium.org/
    description: Production server
paths:
...
  /field-descriptions:
    get:
      operationId: field_descriptions_get
      summary: Returns descriptions for ingest metadata fields. Replacement for field-descriptions.yaml.
      parameters:
        - name: source
          in: query
          required: false
          description: case-insensitive name of the description source (HMFIELD = legacy field-descriptions.yaml; CEDAR = CEDAR)
          schema:
            type: string
            enum:
              - HMFIELD
              - CEDAR
      responses:
        '200':
          description: Array of ingest metadata fields with descriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldDescriptionsResponse'
        '400':
          description: Invalid *source* parameter (i.e., not CEDAR or HMFIELD); invalid parameter
        '404':
          description: No field descriptions (with list of parameters, if specified)
        '5XX':
          description: Unexpected error
  /field-descriptions/{name}:
    get:
      operationId: field_descriptions_name_get
      summary: Returns descriptions for the specified ingest metadata field. Replacement for field-descriptions.yaml.
      parameters:
        - name: name
          in: path
          required: true
          description: name of the metadata field
          schema:
            type: string
            example: ablation_distance_between_shots_x_units
        - name: source
          in: query
          required: false
          description: case-insensitive name of the description source (HMFIELD = legacy field-descriptions.yaml; CEDAR = CEDAR)
          schema:
            type: string
            enum:
              - HMFIELD
              - CEDAR
      responses:
        '200':
          description: Array of ingest metadata fields with descriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldDescriptionsResponse'
        '400':
          description: Invalid *source* parameter (i.e., not CEDAR or HMFIELD); invalid parameter
        '404':
          description: No descriptions (with list of parameters, if specified)
        '5XX':
          description: Unexpected error
  /field-types:
    get:
      operationId: field_types_get
      summary: Return data types for ingest metadata fields. Replacement for field-types.yaml.
      parameters:
        - name: mapping_source
          in: query
          required: false
          description: case-insensitive name of the field-to-type mapping source (HMFIELD = field-types.yaml; CEDAR = CEDAR)
          schema:
            type: string
            enum:
              - HMFIELD
              - CEDAR
        - name: type_source
          in: query
          required: false
          description: case-insensitive name of the type source (HMFIELD = field-types.yaml; XSD = from CEDAR template)
          schema:
            type: string
            enum:
              - HMFIELD
              - XSD
        - name: type
          in: query
          required: false
          description: name of type in either XSD or HMFIELD ontology
          schema:
            type: string
            example: anyURI
      responses:
        '200':
          description: Array of metadata fields with their HMFIELD and/or XSD data types.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldTypesResponse'
        '400':
          description: Invalid value for parameter (e.g., *mapping_source* not HMFIELD or CEDAR); invalid parameter
        '404':
          description: No field-type associations (with list of parameters, if specified)
        '5XX':
          description: Unexpected error
  /field-types/{name}:
    get:
      operationId: field_types_name_get
      summary: Return data types for a specified ingest metadata field. Replacement for field-types.yaml.
      parameters:
        - name: name
          in: path
          required: true
          description: name of the metadata field
          schema:
            type: string
            example: area_unit
        - name: mapping_source
          in: query
          required: false
          description: case-insensitive name of the field-to-type mapping source (HMFIELD = field-types.yaml; CEDAR = CEDAR)
          schema:
            type: string
            enum:
              - HMFIELD
              - CEDAR
        - name: type_source
          in: query
          required: false
          description: case-insensitive name of the type source (HMFIELD = field-types.yaml; XSD = from CEDAR template)
          schema:
            type: string
            enum:
              - HMFIELD
              - XSD
        - name: type
          in: query
          required: false
          description: name of type in either XSD or HMFIELD ontology
          schema:
            type: string
            example: float
      responses:
        '200':
          description: Array of metadata fields with their HMFIELD and/or XSD data types.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldTypesResponse'
        '400':
          description:  Invalid value for parameter (e.g., *mapping_source* not HMFIELD or CEDAR); invalid parameter
        '404':
          description: No field-type associations (with list of parameters, if specified)
        '5XX':
          description: Unexpected error
  /field-types-info:
    get:
      operationId: field_types_indo_get
      summary: Returns a list of all available data types for metadata fields.
      parameters:
        - name: type_source
          in: query
          required: false
          description: case-insensitive name of the type source (HMFIELD = field-types.yaml; XSD = from CEDAR template)
          schema:
            type: string
            enum:
              - HMFIELD
              - XSD
      responses:
        '200':
          description: Array of HMFIELD and/or XSD data types.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldTypesInfoResponse'
        '400':
          description:  Invalid parameter
        '404':
          description: No field types (with list of parameters, if specified)
        '5XX':
          description: Unexpected error
...
  /field-assays:
    get:
      operationId: field_assays_get
      summary: Return associations between ingest metadata fields and the "assays" (dataset data types). Replacement for field-assays.yaml.
      parameters:
        - name: assay_identifier
          in: query
          required: false
          description: the legacy idenetifier for the assay dataset in field_assays.yaml. This may be a data_type, alt-name, or description of the dataset.
          schema:
            type: string
            example: scRNAseq-10xGenomics
        - name: data_type
          in: query
          required: false
          description: the legacy *data_type*--i.e., the key for ingest workflows
          schema:
            type: string
            example: scRNAseq-10xGenomics-v3
        - name: dataset_type
          in: query
          required: false
          description: the *soft assay* dataset type--i.e., the value used by the Rules Engine
          schema:
            type: string
            example: RNASeq
      responses:
        '200':
          description: Array of ingest metadata fields; each field with its associations with assays/dataset types.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldAssaysResponse'
        '400':
          description: Invalid parameter
        '404':
          description: no field to assay relationships found (for the set of parameters, if specified)
        '5XX':
          description: Unexpected error
  /field-assays/{name}:
    get:
      operationId: field_assays_name_get
      summary: Return associations between the specified ingest metadata field and the "assays" (dataset data types). Replacement for field-assays.yaml.
      parameters:
        - name: name
          in: path
          required: true
          description: name of the metadata field
          schema:
            type: string
            example: acquisition_instrument_model
        - name: assay_identifier
          in: query
          required: false
          description: the legacy idenetifier for the assay dataset in field_assays.yaml. This may be a data_type, alt-name, or description of the dataset.
          schema:
            type: string
            example: scRNAseq-10xGenomics
        - name: data_type
          in: query
          required: false
          description: the legacy *data_type*--i.e., the key for ingest workflows
          schema:
            type: string
            example: scRNAseq-10xGenomics-v3
        - name: dataset_type
          in: query
          required: false
          description: the *soft assay* dataset type--i.e., the value used by the Rules Engine
          schema:
            type: string
            example: RNASeq
      responses:
        '200':
          description: Array of ingest metadata fields; each field with its associations with assays/dataset types.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldAssaysResponse'
        '400':
          description: Invalid parameter
        '404':
          description: no field to assay relationships found (for the set of parameters, if specified)
        '5XX':
          description: Unexpected error
  /field-schemas:
    get:
      operationId: field_schemas_get
      summary: Return associations between ingest metadata fields and metadata schemas. Replacement for field-schemas.yaml.
      parameters:
        - name: source
          in: query
          required: false
          description: case-insensitive name of the field-to-type mapping source (HMFIELD = field-types.yaml; CEDAR = CEDAR)
          schema:
            type: string
            enum:
              - HMFIELD
              - CEDAR
        - name: schema
          in: query
          required: false
          description: case-sensitive name of a metadata schema or CEDAR template
          schema:
            type: string
            example: imc
      responses:
        '200':
          description: Associations between ingest metadata fields and metadata schemas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldSchemasResponse'
        '400':
          description: Invalid parameter; invalid value for *source*
        '404':
          description: no field to schema relationships found (for the set of parameters, if specified)
        '5XX':
          description: Unexpected error
  /field-schemas/{name}:
    get:
      operationId: field_schemas_name_get
      summary: Return associations between a specified ingest metadata field and metadata schemas. Replacement for field-schemas.yaml.
      parameters:
        - name: name
          in: path
          required: true
          description: name of the metadata field
          schema:
            type: string
            example: acquisition_instrument_model
        - name: source
          in: query
          required: false
          description: case-insensitive name of the field-to-type mapping source (HMFIELD = field-types.yaml; CEDAR = CEDAR)
          schema:
            type: string
            enum:
              - HMFIELD
              - CEDAR
        - name: schema
          in: query
          required: false
          description: case-sensitive name of a metadata schema or CEDAR template
          schema:
            type: string
            example: imc
      responses:
        '200':
          description: Associations between ingest metadata fields and metadata schemas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldSchemasResponse'
        '400':
          description: Invalid parameter; invalid value for *source*
        '404':
          description: no field to schema relationships found (for the set of parameters, if specified)
        '5XX':
          description: Unexpected error
components:
  schemas:
...
    FieldDescriptionsResponse:
      type: object
      description: Descriptions ingest metadata fields from legacy field-descriptions.yaml
      properties:
        fields:
          type: array
          description: array of fields and their descriptions
          items:
            type: object
            properties:
              code_ids:
                type: array
                description: Codes for the field in different ontologies. HMFIELD codes are from the legacy field_*.yaml files; CEDAR codes are from CEDAR template schemas.
                items:
                  type: string
                  example: ["HMFIELD:1008","CEDAR:9f654d25-4de7-4eda-899b-417f05e5d5c3"]
              name:
                type: string
                description: identifier for metadata field
                example: ablation_distance_between_shots_x_units
              descriptions:
                type: array
                description: descriptions of field, from both HMFIELD and CEDAR
                items:
                  type: object
                  properties:
                    source:
                      type: string
                      description: source of the description
                      example: CEDAR
                    description:
                      type: string
                      description: description string
                      example: Units of x resolution distance between laser ablation shots.
    FieldTypesResponse:
      type: object
      description: Associations between metadata field and data type ingest metadata fields
      properties:
        fields:
          type: array
          description: array of fields and their data types
          items:
            type: object
            properties:
              code_ids:
                type: array
                description: Codes for the field in different ontologies. HMFIELD codes are from the legacy field_*.yaml files; CEDAR codes are from CEDAR template schemas.
                items:
                  type: string
                  example: ["HMFIELD:1018","CEDAR:ea6e26f1-30df-4a8c-9f15-63111ea0b68b"]
              name:
                type: string
                description: identifier for metadata field
                example: area_unit
              types:
                type: array
                description: field type associations
                items:
                  type: object
                  properties:
                    mapping_source:
                      type: string
                      description: source of the field-to-type mapping (i.e., the ontology).
                      example: CEDAR
                    type_source:
                      type: string
                      description: source of the field code (i.e., the type ontology)
                      example: XSD
                    type:
                      type: string
                      description: the term for the type code in the type ontology
                      example: float
    FieldTypesInfoResponse:
      type: object
      description: Details on ingest metadata field types.
      properties:
        types:
          type: array
          description: field types
          items:
            type: object
            properties:
              type_source:
                type: string
                description: source of the field code (i.e., the type ontology)
                example: XSD
              type:
                type: string
                description: the term for the type code in the type ontology
                example: float
...
    FieldAssaysResponse:
      type: object
      description: Associations between fields and "assays" (or datasets generated by assays) for ingest metadata fields
      properties:
        fields:
          type: array
          description: array of fields, each with its associated assays/dataset types
          items:
            type: object
            properties:
              code_ids:
                type: array
                description: Codes for the field in different ontologies. HMFIELD codes are from the legacy field_*.yaml files; CEDAR codes are from CEDAR template schemas.
                items:
                  type: string
                  example: ["HMFIELD:1008","CEDAR:9f654d25-4de7-4eda-899b-417f05e5d5c3"]
              name:
                type: string
                description: name of the metadata field
                example: acquisition_instrument_model
              assays:
                type: array
                description: array of assays/datasets associated with the field
                items:
                  type: object
                  properties:
                    assay_identifier:
                      type: string
                      description: The legacy identifier used in field_assays.yaml. This can be the data_type, an alt-name, or the description of the data_type used in Portal.
                      example: scRNAseq-10xGenomics
                    data_type:
                      type: string
                      description: The legacy data_type--i.e., the key for ingest workflows
                      example: scRNAseq-10xGenomics-v3
                    dataset_type:
                      type: string
                      description: The "soft assay" dataset type--e.g., used by the Rules Engine
                      example: RNASeq
    FieldSchemasResponse:
      type: object
      description: Associations between fields and metadata schemas for ingest metadata fields from legacy field-schemas.yaml
      properties:
        fields:
          type: array
          description: array of fields, each with its associated schemas
          items:
            type: object
            properties:
              code_ids:
                type: array
                description: Codes for the field in different ontologies. HMFIELD codes are from the legacy field_*.yaml files; CEDAR codes are from CEDAR template schemas.
                items:
                  type: string
                  example: ["HMFIELD:1008","CEDAR:9f654d25-4de7-4eda-899b-417f05e5d5c3"]
              name:
                type: string
                description: identifier for metadata field
                example: ablation_distance_between_shots_x_units
              schemas:
                type: array
                description: array of metadata schemas/templates associated with the field
                items:
                  type: object
                  properties:
                    source:
                      type: string
                      description: source of schema mapping
                      example: HMFIELD
                    schema:
                      type: string
                      description: name of schema
                      example: imc3d